+++ src/snyk/snykCode/views/securityIssueTreeProvider.ts
@@ -8,6 +8,7 @@ import { IViewManagerService } from '../../common/services/viewManagerService';
 import { TreeNode } from '../../common/views/treeNode';
 import { IVSCodeLanguages } from '../../common/vscode/languages';
 import { IssueTreeProvider } from './issueTreeProvider';
+import { FEATURE_FLAGS } from '../../common/constants/featureFlags';
 
 export default class CodeSecurityIssueTreeProvider extends IssueTreeProvider {
   constructor(
@@ -38,9 +39,19 @@ export default class CodeSecurityIssueTreeProvider extends IssueTreeProvider {
     return `${dir} - ${issueCount} ${issueCount === 1 ? 'vulnerability' : 'vulnerabilities'}`;
   }
 
-  protected getIssueFoundText(nIssues: number): string {
+  protected getIssueFoundText(nIssues: number, ignoredIssueCount: number): string {
     if (nIssues > 0) {
-      return nIssues === 1 ? `${nIssues} vulnerability found by Snyk` : `✋ ${nIssues} vulnerabilities found by Snyk`;
+      let text;
+      if (nIssues === 1) {
+        text = `${nIssues} vulnerability found by Snyk`;
+      } else {
+        text = `✋ ${nIssues} vulnerabilities found by Snyk`;
+      }
+      const isIgnoresEnabled = configuration.getFeatureFlag(FEATURE_FLAGS.consistentIgnores);
+      if (isIgnoresEnabled) {
+        text += `, ${ignoredIssueCount} ignored`;
+      }
+      return text;
     } else {
       return '✅ Congrats! No vulnerabilities found!';
     }