+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.9.2]
 - Injects custom styling for the HTML panel used by Snyk Code for consistent ignores.
+- Add warning messages in the Tree View for the issue view options used in consistent ignores.
 
 ## [2.8.1]
 - Lower the strictness of custom endpoint regex validation so that single tenant APIs are allowed.